Introduce commons-lang3 to partially replace guava usage
Not all developers are comfortable pulling in such a large dependency as
guava. This library really doesn't need or warrant such a big
dependency. This patch introduces commons-lang3 as a new dependency and
partially replaces our guava usage with it. The main focus here is on
Precondition checking in various places.