Shift host and maxBodySize selection into ClientCredential
Every other language binding library for EdgeGrid signing reads the
maxBodySize and host out of the client credential object. This patch
adjust the code to add it there, and modifies the signing binding
infrastrucutre to use that host instead of whatever host was
pre-selected. In real terms this means we replace the "host" header if
it exists, and when possible we modify the request directly to change
the hostname there. This does not work for the RestAssured variant
because of how its request initialization works.