Introducing ClientCredentialProvider
This patch introduces ClientCredentialProvider, which can be implemented
by a user to select a good ClientCredential for the request at signing
time. A reference to the Request intermediart object is provided to
facilitate ClientCredential selection.

All of the signers will also accept a simple ClientCredential, which
transparently instantiates a DefaultClientCredentialProvider with that
object and subsequently uses the same ClientCredential for every
request.