Support URI fragments in Request
We only really care about the path and query string out of a URI for the
purpose of generating a signature. We care about the scheme too, but in
reality it's always going to be HTTPS.

This patch alters the Request object so its field for this is simply
"uri", and the method to set one (also in the builder) is just "#uri()".
We also accept a String now for convenience (parsing it with
URI#create()) and we just pluck out the path and query string. All other
fields are explicitly set to null.

I have left a deprecated wrapper method with the prior builder method
name.