Produce OSGi bundle headings in manifest
Some of our users are using OSGi bundles in their application, and our
library doesn't play nicely in such an environment. It seems we just
needed to use make some changes to how we produce the final MANIFEST.MF
file. This patch pulls in the maven-bundle-plugin to product an
augmented MANIFEST.MF file, and updates the maven-jar-plugin to use it.