Fix #35
A user reported that the build was failing for him, and I quickly
discovered that the failure was present on Java 11 but not Java 8. After
some digging, I realized that the issue had something to do with not
understanding or misusing the constructors for URI. I addressed this by
changing the test to use URI#create().

In the process of fixing this, I realized that we would not properly
handle paths that contain percent-encoded characters. We seem to need to
use URI#getRawPath() for the path, but URI#getQuery() (not
URI#getRawQuery()) for the query string. I've fixed the signer to do
this correctly, and added another unit test to cover it.